+++ ui/src/App.tsx
@@ -10,7 +10,7 @@ import InboxIcon from "@mui/icons-material/MoveToInbox";
 import ListItem from "@mui/material/ListItem";
 import ListItemIcon from "@mui/material/ListItemIcon";
 import ListItemText from "@mui/material/ListItemText";
-import { Breadcrumbs } from "./components/breadcrumbs/Breadcrumbs";
+import { Breadcrumbs } from "./components/common/Breadcrumbs";
 import { Namespaces } from "./components/namespaces/Namespaces";
 import { Pipeline } from "./components/pipeline/Pipeline";
 import logo from "./images/icon-on-blue-bg.png";
