--- ui/src/components/breadcrumbs/Breadcrumbs.css
@@ -1,12 +1,0 @@
-.Breadcrumbs {
-    margin-bottom: 15px !important;
-}
-
-.Breadcrumbs-link {
-    text-decoration: none;
-    color: inherit;
-}
-
-.Breadcrumbs-link:hover {
-    text-decoration: underline;
-}
