+++ ui/src/components/common/Breadcrumbs/index.tsx
@@ -1,0 +1,65 @@
+import React, { useMemo } from "react";
+import { useLocation, Link } from "react-router-dom";
+import MUIBreadcrumbs from "@mui/material/Breadcrumbs";
+import Typography from "@mui/material/Typography";
+
+import "./style.css";
+
+export function Breadcrumbs() {
+  const location = useLocation();
+
+  const crumbs = useMemo(() => {
+    const { pathname } = location;
+    if (pathname === "/") {
+      // Namespaces (home)
+      return (
+        <Typography
+          data-testid="namespace-breadcrumb"
+          className="Breadcrumbs-typ"
+          sx={{ fontWeight: 600 }}
+        >
+          Namespaces
+        </Typography>
+      );
+    }
+    const pathParts = pathname.split("/");
+    switch (pathParts.length) {
+      case 5: // pipeline view
+        return [
+          <Link key={"pipeline-view"} to="/" className="Breadcrumbs-link">
+            Namespaces
+          </Link>,
+          <Typography
+            key={"pipeline-typ"}
+            data-testid="pipeline-breadcrumb"
+            className="Breadcrumbs-typ"
+            sx={{ fontWeight: 600 }}
+          >
+            {`${pathParts[2]} (${pathParts[4]})`}
+          </Typography>,
+        ];
+      default:
+        // Unsupported path
+        return (
+          <Typography
+            data-testid="unknown-breadcrumb"
+            className="Breadcrumbs-typ"
+            sx={{ fontWeight: 600 }}
+          >
+            Unknown
+          </Typography>
+        );
+    }
+  }, [location]);
+
+  return (
+    <MUIBreadcrumbs
+      data-testid="mui-breadcrumbs"
+      className="Breadcrumbs"
+      separator="›"
+      aria-label="breadcrumb"
+    >
+      {crumbs}
+    </MUIBreadcrumbs>
+  );
+}
